requirements.txt
aiogram>=3.15.0
python-dotenv>=1.0.0
google-genai>=1.0.0
redis>=5.0.0